pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ikv-secrets"
version = "0.1.0"
description = "Secure environment variable access for IronKeyVault"
readme = "README.md"
license = "MIT"
requires-python = ">=3.9"
authors = [
    { name = "IronKeyVault", email = "support@ironkeyvault.com" }
]
keywords = ["secrets", "environment", "vault", "security", "ironkeyvault"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "httpx>=0.25.0",
    "keyring>=24.0.0",
    "pyyaml>=6.0",
    "click>=8.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[project.scripts]
ikv-secrets = "ikv_secrets.cli:main"

[project.urls]
Homepage = "https://github.com/IronKeyVault/ikv-secrets"
Documentation = "https://ironkeyvault.com/docs/sdk"
Repository = "https://github.com/IronKeyVault/ikv-secrets"
Issues = "https://github.com/IronKeyVault/ikv-secrets/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/ikv_secrets"]

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.mypy]
python_version = "3.9"
strict = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --cov=ikv_secrets"
